requirements.txt
djangorestframework==3.14.0
django==4.2.5
requests==2.31.0
drf-spectacular==0.26.4